Improve validity of models

+++ V2022/Location.cs
@@ -5,9 +5,9 @@
         public string DataCenter { get; set; }
         public string World { get; set; }
         public string District { get; set; }
-        public int Ward { get; set; }
-        public int Plot { get; set; }
-        public int Apartment { get; set; }
+        public ushort Ward { get; set; }
+        public ushort Plot { get; set; }
+        public ushort Apartment { get; set; }
         public bool Subdivision { get; set; }
         public string Shard { get; set; }
     }
